Extract general argument types to new module
The ToArgs class and utility functions for creating arguments have
been moved to a new module, allowing modules to be created for
different docker commands.
